+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to secrets-store-csi-driver-provider-gcp will be documented 
 
 * The `-write_secrets` flag defaults to `false`. This requires `v0.0.21+` of the `secrets-store-csi-driver`. [#98](https://github.com/GoogleCloudPlatform/secrets-store-csi-driver-provider-gcp/issues/98)
 
+### Added
+
+* Specify `auth: provider-adc` in the SecretProviderClass to use application default credentials instead of the mount's pod's workload identity. [#101](https://github.com/GoogleCloudPlatform/secrets-store-csi-driver-provider-gcp/issues/101)
+
 ## v0.4.0
 
 Images: